requirements.txt
Django>=4.2,<5.0

# Postgress connection with django client library
psycopg2-binary>=2.9.6

# Requesting library
requests>=2.30.0

# Redis, Caching
django-redis>=5.2.0
redis>=4.5.5

# Celery
celery>=5.3.0

# Telegram
python-telegram-bot>=20.3

#Django Rest Framework
djangorestframework>=3.14.0